backend/python/requirements.txt
Django==4.1.2
PyMySQL==1.0.2
mysqlclient==2.1.1

django-environ==0.9.0
django-widget-tweaks==1.4.12

selenium==4.4.3
beautifulsoup4==4.11.1
chromedriver-binary==105.0.5195.19.0

djangorestframework==3.13.1
django-cors-headers==3.13.0

django-debug-toolbar==3.6.0

psutil==5.9.1
factory_boy==3.2.1

matplotlib==3.6.1